fix: match CTA text width to Figma + load Geist in Storybook (#64)
The text column capped its max-width on the padded wrapper, shrinking the
copy to ~298px and over-wrapping the heading/paragraph. Move the padding
outside the 462px cap so wrapping matches the Figma frame, and bump the
blue panel's right radius closer to the design.

Also load Geist in the Storybook preview (as the app does in layout.tsx)
so stories render with the real font metrics instead of a fallback.
